Add files via upload
Happy to share that I've completed the fifth task of my internship at #ProdigyInfoTech. This task was to create a Weather App Web Application. Using HTML to structure, CSS to style the application, coding in the notepad editor and JavaScript to add interactivity between the user to know about the Weather condition entered by the user which will lead the  user to know about the temperature, humidity and windspeed . #webDevelopment #Internship #ProdigyInfoTech